0-AURA_agent/main_agent/sub_agents/financial_behavior_agent/requirements.txt
google-cloud-firestore
google-generativeai
google-adk==0.1.0
google-cloud-aiplatform
python-dotenv
pandas